% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_assessments.R
\name{add_hydrolic}
\alias{add_hydrolic}
\title{Maximal acceptable impervious area based on hydrolic conditions}
\usage{
add_hydrolic(site_data, max_area_table, q_rain)
}
\arguments{
\item{site_data}{The site specific data loaded with function "loda_site_data"}

\item{max_area_table}{Table created with}

\item{q_rain}{characteristic rainfall in L/(s*ha)}
}
\value{
the combined max_area table is extend by a row with the result of the
hydolic assessment.
}
\description{
This function calculates the maximal acceptabale impervious area
}
