% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_assessments.R
\name{add_max_areas}
\alias{add_max_areas}
\title{Maximal acceptable impervious area per substance}
\usage{
add_max_areas(
  combined_concentration_table,
  site_data,
  q_rain,
  t_rain = rain[1] * 60
)
}
\arguments{
\item{combined_concentration_table}{The table created by function
"combine_concentration_tables"}

\item{site_data}{The site specific data loaded with function "load_site_data"}

\item{q_rain}{Average rain intensity in L/(s*ha)}

\item{t_rain}{Duration of the rain event in s}
}
\value{
the combined concentration table is extend by the columns
max_area_catch_ha: the maximal tolerable impervious area in the whole catchment
max_area_plan_ha: the corresponding area in the planning area
max_area_plan_percent: the Percantage of the currently impervious area.
Values < 100 means a reduction of impervious area is necessary
}
\description{
This function calculates the maximal acceptabale impervious
area per substance depending on the threshold value type either by dynamic
or steady state concentration modeling
}
