% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{area_from_load}
\alias{area_from_load}
\title{Calculate pollutant input from runoff area area within a rain event}
\usage{
area_from_load(load_runoff, Ci_storm, coeff_runoff, q_rain, t_rain)
}
\arguments{
\item{load_runoff}{Pollutant load from urban area. Mass unit is one unit
larger (factor 1000) as in concentration. For example: if Ci_storm is in
ug/L, load must be in mg.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i.}

\item{coeff_runoff}{runoff coefficient of runoff area.}

\item{q_rain}{rain amount in mm/(ha*s)}

\item{t_rain}{duration of rain in s}
}
\value{
Numeric value of the according connectable area in ha
}
\description{
Calculate pollutant input from runoff area area within a rain event
}
