% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{assess_all_hazards}
\alias{assess_all_hazards}
\title{Runs function \code{\link[=immission_assessment]{immission_assessment()}} for all substances that might
pose a risk and returns the results in three tables}
\usage{
assess_all_hazards(
  hazard_list,
  site_data,
  c_table,
  q_rain,
  t_rain,
  c_type = "average"
)
}
\arguments{
\item{hazard_list}{Hazard list created by function \code{\link[=check_all_substances]{check_all_substances()}}}

\item{site_data}{Site data list created by function \code{\link[=load_site_data]{load_site_data()}}}

\item{c_table}{Table with concentions in the river, in rainwater runoff and
threshold values}

\item{q_rain}{Intensity of a rain event in L/(ha*s)}

\item{t_rain}{Length of a rain event in s}

\item{c_type}{Character value specifiyng the type of concentration that
is used for the assessment. Either "average" for median value or "worstcase"
for 95th quantile concentration in rainwater runoff.}
}
\value{
A list of three tables. 1) General information about the whole
urbanised area calculated with data of the landuse types.
2) Detailad information about connectable area of the planning area taking
into account the status quo of the surrounding urbanised area. 3) Detailad
information about the connectable area of the planning area, scaled down from
the overall connectable area.
}
\description{
Runs function \code{\link[=immission_assessment]{immission_assessment()}} for all substances that might
pose a risk and returns the results in three tables
}
