% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{assess_discharges}
\alias{assess_discharges}
\title{assess_discharges}
\usage{
assess_discharges()
}
\value{
ratio between tolerable and relevant discharge
}
\description{
assess_discharges
}
\examples{
assess_discharges()

}
