% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{calculate_relevant_discharge}
\alias{calculate_relevant_discharge}
\title{calculate_relevant_discharge}
\usage{
calculate_relevant_discharge(
  data = r2q::import_hydrology_data(),
  area_classification = T,
  path_to_area = system.file("extdata/areas_hydrology.csv", package = "r2q")
)
}
\arguments{
\item{data}{hydrology (tibble as retrieved by \code{\link{import_hydrology_data}})}

\item{area_classification}{if TRUE perfomes area specific evaluation based on table}

\item{path_to_area}{path the csv file containin garea information}
}
\value{
returns relevant discharge from planning area
}
\description{
calculate_relevant_discharge
}
\examples{

}
