% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{calculate_tolerable_discharge}
\alias{calculate_tolerable_discharge}
\title{Title}
\usage{
calculate_tolerable_discharge(hydrology = hydrology_data, verbose = TRUE)
}
\arguments{
\item{hydrology}{}

\item{verbose}{if TRUE returns results as informative messages, If FALSE only return numeric value for planning area.}
}
\value{
returns tolerable discharge for the planning area based on the data in hydrology table
}
\description{
Title
}
\examples{
 
calculate_tolerable_discharge(hydrology = hydrology_data, verbose = F)
16.45757
calculate_tolerable_discharge(hydrology = hydrology_data, verbose = T)
"Based on provided input data a tolerable annual discharge flow of 283.036 L/s was calculated for Aba. For the planning area this corresponds to 16.46 L/s" 
}
