% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{check_all_substances}
\alias{check_all_substances}
\title{Check if substances pose a risk to the surface water}
\usage{
check_all_substances(c_table, c_type = "average")
}
\arguments{
\item{c_table}{Table of concentration (rainwater, river, threshold value)}

\item{c_type}{Character value specifiyng the type of concentration that
is used for the assessment. Either "average" for median value or "worstcase"
for 95th quantile concentration in rainwater runoff.}
}
\value{
A list with all substances defined in c_table and assigning either 1
or TRUE, if the substance might pose a risk, "Inf" if the substance does not
constrain the seperate sewer connection, and "-Inf" if ther substance should
not be discharged into the surface water at all, because concentration is
already too high.
}
\description{
Check if substances pose a risk to the surface water
}
