% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{check_pollutant_impact}
\alias{check_pollutant_impact}
\title{Check Pollutant Impact}
\usage{
check_pollutant_impact(Ci_river, Ci_threshold, Ci_storm, pollutant_name = "")
}
\arguments{
\item{Ci_river}{Background concentration for substance i. Concentration unit
must fit to Ci_threshold and Ci_storm.}

\item{Ci_threshold}{Threshold value for substance i. Concentration unit
must fit to Ci_river and Ci_storm.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i. Concentration unit
must fit to Ci_threshold and Ci_river.}

\item{pollutant_name}{String defining the name of the Pollutant}
}
\value{
Inf if the pollutant is no constraint, -Inf if the pollutant should not be
discharged at all, and TRUE if the tolerable load can be calculated
}
\description{
Checks if the pollutant i is a constraint for the connected area
}
