% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{combine_concentration_tables}
\alias{combine_concentration_tables}
\title{Combine the three R2Q concentration tables
(treshold values, background and stromwater)}
\usage{
combine_concentration_tables(
  threshold_table,
  storm_table,
  background_table,
  onlyComplete = TRUE
)
}
\arguments{
\item{threshold_table}{the internal threshold table
loaded with "get_thresholds"}

\item{storm_table}{the internal stormwater concentration table
loaded with "get_stormwater_concentrations"}

\item{background_table}{the filled in background pollution data table
(Excel File) loaded with "load_background_data"}

\item{onlyComplete}{If TRUE (default) only such substances remain in the
data frame with complete concentration triplet (threshold, stromwater, river)
in the same unit}
}
\value{
A data frame with the information threshold values, threshold value type,
background concentration and stormwater concentration per substance and
concentration unit.
}
\description{
This function reduces the input tables to the columns "Substance", "Unit",
renames the Value column according to the table data and combines all tables
}
\details{
Caution: This function uses the Column names of the tables. Do not change
the first two column names of the pollution data. Column names must be
\enumerate{
\item "Substance" and 2) "Unit".
}
}
