% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology_dev.R
\name{criteria_test}
\alias{criteria_test}
\title{Checks whether hydrologic criteria are fullfilled}
\usage{
criteria_test(Q_zulaessig, Q_ist)
}
\arguments{
\item{Q_zulaessig}{Q_zulaessig}

\item{Q_ist}{Q_ist}
}
\value{
????
}
\description{
Checks whether hydrologic criteria are fullfilled
}
