% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_allowed_area}
\alias{get_allowed_area}
\title{Get allowed impervious area}
\usage{
get_allowed_area(f_D, Q_tol, q_rain)
}
\arguments{
\item{Q_tol}{Tolearble discharge into the surface water in L/s}

\item{q_rain}{presipitaion rate in L/(s * ha)}

\item{f_DA}{Run-off coefficient of impervious area}
}
\value{
allowed impervious area in ha
}
\description{
Get allowed impervious area
}
