% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_areaType_runoff}
\alias{get_areaType_runoff}
\title{This function loads the area type specific pollutant runoff concentration
obtained by the OgRe Dataset and multiplies it with the proportion of the
correspoding area type in the catchment.}
\usage{
get_areaType_runoff(areaType_vector = c(0.4, 0.4, 0.2, 0))
}
\arguments{
\item{areaType_vector}{A Vector of proportions for residential_suburban,
residential_city, industry, street area types. The sum of the vector should
be 1}
}
\value{
A dataframe with the columns "Substance", "unit", "Mean" which is the
median value and "Q95" which is the 95th quantile.
}
\description{
This function loads the area type specific pollutant runoff concentration
obtained by the OgRe Dataset and multiplies it with the proportion of the
correspoding area type in the catchment.
}
