% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology_dev.R
\name{get_average_runoff_coef}
\alias{get_average_runoff_coef}
\title{Get average runoff coefficient}
\usage{
get_average_runoff_coef(A_ba, q_zul, R_Spende)
}
\arguments{
\item{A_ba}{A_ba}

\item{q_zul}{q_zul}

\item{R_Spende}{R_Spende}
}
\value{
???
}
\description{
Get average runoff coefficient
}
