% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_background.R
\name{get_backgrounds}
\alias{get_backgrounds}
\title{get background concentrations for SUW before rain events}
\usage{
get_backgrounds(data.dir, SUW_type = "river", substances)
}
\arguments{
\item{data.dir}{path where background tables are stored}

\item{SUW_type}{lake or river, river is used as default}

\item{substances}{array of substances}
}
\value{
data.frame with background concentrations
}
\description{
loads background concentrations
based on entry by user or default values
from csv table
}
