% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_default_background}
\alias{get_default_background}
\title{get background concentrations for SUW before rain events}
\usage{
get_default_background(SUW_type = "river")
}
\arguments{
\item{SUW_type}{"lake" or "river", "river" is used as default}
}
\value{
data.frame with background concentrations
}
\description{
loads background concentrations
based on entry by user or default values
from csv table
}
