% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{get_functionsID}
\alias{get_functionsID}
\title{Loads the table with function IDs}
\usage{
get_functionsID()
}
\value{
data.frame with function IDs and additional 1 to 3 characterizations
}
\description{
Loads the table with function IDs
}
