% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology_dev.R
\name{get_Hq1_pnat}
\alias{get_Hq1_pnat}
\title{Calculate natural runoff based on slope of landscape}
\usage{
get_Hq1_pnat(gefaelle = 0.1)
}
\arguments{
\item{gefaelle}{gefaelle (unit ??)}
}
\value{
?????
}
\description{
Calculate natural runoff based on slope of landscape
}
\examples{
get_Hq1_pnat(gefaelle = 0.1)
get_Hq1_pnat(gefaelle = 0.5)
get_Hq1_pnat(gefaelle = 2)
get_Hq1_pnat(gefaelle = 50)
}
