% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_rain.R
\name{get_HQ_time_interval}
\alias{get_HQ_time_interval}
\title{get_HQ_time_intterval}
\usage{
get_HQ_time_interval(
  area_catch,
  river_cross_section,
  river_length,
  Hq_pnat1_catch = NULL,
  slope = 0.1
)
}
\arguments{
\item{area_catch}{The catchment area in km2}

\item{river_cross_section}{The average river cross section in the catchment
in m2}

\item{river_length}{The length of the affected urban river stretch in m}

\item{Hq_pnat1_catch}{the natural catchment discharge for a yearly rain event
in L/(s*km2). If NULL it will be estimated by slope and area of the catchment}

\item{slope}{Average slope of the catchment in \% (Default is 0.1)}
}
\value{
Travel time of discharged water within the natural catchment in minutes
}
\description{
This function calculates the time in minutes that is needed for the water to
travel through the affected urban river stretch for a yearly rain event,
based on a natural catchment discharge
}
\examples{
get_HQ_time_interval(
area_catch = 5.62,
river_cross_section = 0.54, 
river_length = 5000)

}
