% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSTRA_timeseries.R
\name{get_KOSTRA}
\alias{get_KOSTRA}
\title{Get KOSTRA rain characteristics}
\usage{
get_KOSTRA(coord_vector, duration_string, location_name = NULL, plot = TRUE)
}
\arguments{
\item{coord_vector}{coordinates in ETRS89. See Datails for more information.}

\item{duration_string}{Duration of precipitation in minutes}

\item{location_name}{used for plot title. Default is NULL}

\item{plot}{boolean (TRUE for plotting or FALSE if plotting is not required)}
}
\value{
A list with elements "plot" and "data" (numeric results)
and also produce a ggplot2 plot if parameter plot = TRUE)
}
\description{
Get KOSTRA rain characteristics
}
\details{
The KOSTRA Data is available for a grid of x x x km. The location
is given in coordinates in the ETRS89 system
(For information see: https://epsg.io/3034)
Longitudes and Lattidudes in WGS84 can be converted into ETRS89 here:
https://epsg.io/transform#s_srs=4326&t_srs=3034
}
\examples{
  
# Example default values Herne
herne <- get_KOSTRA()
herne$plot
herne$data
  
# Example 10 min for Berlin
berlin <- get_KOSTRA(coord_vector = c(4217676.98, 2862423.69),
duration_string = "0010", location_name = "Berlin", plot = TRUE)
berlin$plot
berlin$data

}
