% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_landuse_runoff}
\alias{get_landuse_runoff}
\title{Loads landuse specific pollutant runoff concentration obtained by the OgRe
Dataset}
\usage{
get_landuse_runoff()
}
\value{
A data frame with the columns "Substance", "unit",  median  and 95th quantile
of landuses "residential_suburban", "residential_city", "commercial" and
"main_road"
}
\description{
Loads landuse specific pollutant runoff concentration obtained by the OgRe
Dataset
}
