% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planning_details.R
\name{get_planningLoad}
\alias{get_planningLoad}
\title{Load of one parameter from one specific surface}
\usage{
get_planningLoad(
  planning_data,
  sID,
  fID,
  q_rain,
  t_rain,
  y_rain,
  thresholdTable,
  function_c_table = NULL
)
}
\arguments{
\item{planning_data}{The Excel sheet "planning_area_details" loaded by
\code{\link[=load_planning_details]{load_planning_details()}}}

\item{sID}{Substance ID as defined in the package substance ID table
(see \code{\link[=get_subID]{get_subID()}})}

\item{fID}{Area function ID as defined in the package functionID table
(see \code{\link[=get_functionsID]{get_functionsID()}})}

\item{q_rain}{Rain intensity in L/(ha*s)}

\item{t_rain}{Rain length in s}

\item{y_rain}{Yearly rain amount in mm}

\item{thresholdTable}{Table of threshold values. Can be loaded with
\code{\link[=get_thresholds]{get_thresholds()}}.}

\item{function_c_table}{A table of surface specific runoff concentrations
from the package. If NULL it is loaded automatically within the function.}
}
\value{
Pollutant load per event or per year, depending on the treshold value
definition. The mass unit is either mg or ug, depending on the input
concentration unit.
}
\description{
Load of one parameter from one specific surface
}
