% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_q_area_partial}
\alias{get_q_area_partial}
\title{Calculate relevant discharge to surface water (rainwater only)}
\usage{
get_q_area_partial(A_ba, f_DA, R_spende = 1)
}
\arguments{
\item{A_ba}{A_ba}

\item{f_DA}{f_DA}

\item{R_spende}{R_spende}
}
\value{
???
}
\description{
Calculate relevant discharge to surface water (rainwater only)
}
