% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_q_area_total}
\alias{get_q_area_total}
\title{Calculate relevant discharge of table of areas}
\usage{
get_q_area_total(dataframe, R_Spende = 1)
}
\arguments{
\item{dataframe}{dataframe}

\item{R_Spende}{R_Spende}
}
\value{
???
}
\description{
Calculate relevant discharge of table of areas
}
