% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_q_max}
\alias{get_q_max}
\title{Calculates tolerable hydraulic burden based on natural runoff estimation}
\usage{
get_q_max(Hq1pnat_catch, x = 0.1, area_con_catch, area_catch)
}
\arguments{
\item{Hq1pnat_catch}{natural discharge of cathcment area (area_catch)
in L/(s*km²)}

\item{x}{default: r2q::get_x()}

\item{area_con_catch}{connected area of planning area in km²}

\item{area_catch}{catchment area until point of discharge in km²}
}
\value{
tolerable discharged cumulative flow of planning area in m³/s
}
\description{
Calculates tolerable hydraulic burden based on natural runoff estimation
}
