% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{get_siteInfoID}
\alias{get_siteInfoID}
\title{Load the table with site data variable IDs}
\usage{
get_siteInfoID()
}
\value{
data.frame with site specific variables IDs, variable names and units
}
\description{
Load the table with site data variable IDs
}
