% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_spec_runoff}
\alias{get_spec_runoff}
\title{This function loads the landuse specific pollutant runoff concentration
obtained by the OgRe Dataset and multiplies it with the proportion of the
correspoding area type in the catchment.}
\usage{
get_spec_runoff()
}
\value{
A dataframe with the columns "Substance", "unit", "Mean" which is the
median value and "Q95" which is the 95th quantile.
}
\description{
This function loads the landuse specific pollutant runoff concentration
obtained by the OgRe Dataset and multiplies it with the proportion of the
correspoding area type in the catchment.
}
