% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stormwater_concentrations.R
\name{get_stormwater_concentrations}
\alias{get_stormwater_concentrations}
\title{get average concentrations in stormwater runoff}
\usage{
get_stormwater_concentrations(data.dir, substances)
}
\arguments{
\item{data.dir}{path where OgRe_data are stored}

\item{substances}{substance name or vector of substance names in German}
}
\value{
data.frame with mean(x), mean (log x) and sd (log x)
}
\description{
assembles concentrations in stormwater runoff
for selected substances bases on Berlin OgRe data set
from csv table. IN addition to averages, coeffs for log-normal
distribution are assessed
}
