% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_stormwaterRunoff}
\alias{get_stormwaterRunoff}
\title{Load landuse specific pollutant runoff concentration}
\usage{
get_stormwaterRunoff(
  runoff_effective_mix = list(c(40, 40, 20, 0), c(20, 40, 20, 20)),
  mix_names = c("is", "pot")
)
}
\arguments{
\item{runoff_effective_mix}{List of numeric vectors. Each vector must contain
4 values representing the areal proportion of "residential suburban",
"residential city", "commercial" and "main road" landuse types in percent.}

\item{mix_names}{A character vector with names for each landuse mix}
}
\value{
A dataframe with the columns "Substance", "unit", the median and 95th
quantile concentrations of all four landuse types and for the defined landuse
combinations
}
\description{
obtained by the OgRe Dataset and multiplies it with the proportion of the
correspoding area type in the catchment.
}
