% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{get_subID}
\alias{get_subID}
\title{Load the table with substance IDs}
\usage{
get_subID()
}
\value{
data.frame with substance IDs, substance names within the OgRe-
data set, clean substance names, substance unit
and substance groups in english and german
}
\description{
Load the table with substance IDs
}
