% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_x}
\alias{get_x}
\title{Calculate acceptable additional runoff factor x}
\usage{
get_x(Hq1_pnat = NA, Hq2_pnat = NA)
}
\arguments{
\item{Hq1_pnat}{Hq1_pnat ???}

\item{Hq2_pnat}{Hq2_pnat ???}
}
\value{
???
}
\description{
Calculate acceptable additional runoff factor x
}
