% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{hydrology_assessment}
\alias{hydrology_assessment}
\title{Maximal connectable impervious area based on hydrologic conditions}
\usage{
hydrology_assessment(site_data, q_rain)
}
\arguments{
\item{site_data}{The site specific data loaded with function "loda_site_data"}

\item{q_rain}{characteristic rainfall in L/(s*ha)}
}
\value{
the combined max_area table is extend by a row with the result of the
hydolic assessment.
}
\description{
Maximal connectable impervious area based on hydrologic conditions
}
\details{
If the planning area is identical with the urban area lines 3, 4 and 7 are
not valid. Line 8 is the required throttel for both, the urban area and the
planning area.
The reason for this is: for the planning area "no_runoff" landuse considered
for discharge calculation (-> possible runoff areas in the future), while
this is not the case for the urban area (-> Status quo assessment). If the
planning area is no "real" planning area, that would not make any sense.
}
