% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{immission_assessment}
\alias{immission_assessment}
\title{Automated immission assessment}
\usage{
immission_assessment(
  site_data,
  c_table,
  q_rain,
  t_rain,
  substance,
  hazard_list,
  c_type = "average"
)
}
\arguments{
\item{site_data}{Site data list created by function \code{\link[=load_site_data]{load_site_data()}}}

\item{c_table}{Table with concentions in the river, in rainwater runoff and
threshold values}

\item{q_rain}{Intensity of a rain event in L/(ha*s)}

\item{t_rain}{Length of a rain event in s}

\item{substance}{Substance name (as definied in c_table)}

\item{hazard_list}{Hazard list created by function \code{\link[=check_all_substances]{check_all_substances()}}}

\item{c_type}{Character value specifiyng the type of concentration that
is used for the assessment. Either "average" for median value or "worstcase"
for 95th quantile concentration in rainwater runoff.}
}
\value{
List with all R2Q Immission assessment output value for the substance
}
\description{
This functions works with a site data list created by function
\code{\link[=load_site_data]{load_site_data()}}, with a concentration table (see details), and with
a hazard list created by function \code{\link[=check_all_substances]{check_all_substances()}}
}
\details{
Something about the c_table
}
