% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{import_hydology_data}
\alias{import_hydology_data}
\title{Import hydrology data}
\usage{
import_hydology_data(
  csv_file = system.file("extdata/hydrology.csv", package = "r2q")
)
}
\arguments{
\item{csv_file}{path to csv file (default: system.file("inst/extdata/hydrology.csv",
package = "r2q"))}
}
\value{
hydrology data
}
\description{
Import hydrology data
}
\examples{
import_hydology_data() 
}
