% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{Input_event}
\alias{Input_event}
\title{Calculate pollutant input from runoff area area within a rain event}
\usage{
Input_event(area_runoff, Ci_storm, coeff_runoff, q_rain, t_rain)
}
\arguments{
\item{area_runoff}{Connected runoff area in ha.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i.}

\item{coeff_runoff}{runoff coefficient of runoff area.}

\item{q_rain}{rain amount in L/(ha*s)}

\item{t_rain}{duration of rain in s}
}
\value{
maximal pollutant input in mass per rain event. The mass unit depends
on the runoff concentration mass unit (one unit larger: factor 1000, i.e. if
concentration is in ug/L, the pollutant load is in mg/event)
}
\description{
Calculate pollutant input from runoff area area within a rain event
}
