% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{KOSTRA_string_from_numeric}
\alias{KOSTRA_string_from_numeric}
\title{KOSTRA_string_from_numeric}
\usage{
KOSTRA_string_from_numeric(time_numeric)
}
\arguments{
\item{time_numeric}{numeric value inline with KOSTRA duration}
}
\value{
returns string suitable for input in get KOSTRA function
}
\description{
KOSTRA_string_from_numeric
}
\examples{
KOSTRA_string_from_numeric(10)
KOSTRA_string_from_numeric(360)

}
