% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_rain.R
\name{lin_interpolation}
\alias{lin_interpolation}
\title{lin_interpolation}
\usage{
lin_interpolation(x1, x2, y1, y2, x_is)
}
\arguments{
\item{x1}{x value of first data point}

\item{x2}{x value of second data point}

\item{y1}{y value of first data point}

\item{y2}{y value of second data point}

\item{x_is}{corresponding x value to the searched y value}
}
\value{
Y-Value to the corresponding x value in the unit of the other y values
}
\description{
Linear interpolation between two data poins
}
\examples{
lin_interpolation(x1 = 60, x2 = 90, y1 = 30, y2 = 55, x_is = 70)

}
