% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_areaTypes}
\alias{load_areaTypes}
\title{Loading all details about catchment area types}
\usage{
load_areaTypes(
  data.dir = NULL,
  filename,
  residential_city = c(0.75, 100/3, 100),
  residential_suburban = c(0.75, 100/3, 100),
  commercial = c(0.75, 100/3, 100)
)
}
\arguments{
\item{data.dir}{The directory of the entry data table.}

\item{filename}{Name of the R2Q-Excel File including ".xlsx".}

\item{residential_suburban, residential_city, commercial}{vectors of 3
containg 1) fD value of the landuse type, 2) the proportion of the landuse
type within the catchment area in percent and 3) the proportion of area
connected to a seperate sewer systen within the landuse type in percent.}
}
\value{
A vector of length 5. Entries 1 to 4 describe the proportion of the area
types "residential_suburban", "residential_city", "industry" and (high-
traffic) "street". The proportion is referred only to the connected area.
The 5th value is the overall proportion of connected area.
}
\description{
this functions loads the data from the sheet "surface_areaType" within the data
entry excel file
}
\examples{
load_site_data(
data.dir = system.file("extdata/Data_entry", package = "r2q"), 
filename = "Baukau.xlsx")

}
