% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_background_data}
\alias{load_background_data}
\title{Loading local background concentration}
\usage{
load_background_data(
  data.dir,
  filename,
  default_for_na = TRUE,
  SUW_type = "river"
)
}
\arguments{
\item{data.dir}{The directory of the entry data table.}

\item{filename}{Name of the R2Q-Excel File including ".xlsx".}

\item{default_for_na}{If TRUE, default values are used for substances that
were not measured}

\item{SUW_type}{Only used if default_for_na is TRUE. "lake" or "river",
"river" is used as default}
}
\value{
A data frame background concentration as defined in the Excel sheet. If
default values are used this is documented in the "comment" column.
}
\description{
This functions loads the data from the sheet "pollution_data" within the
R2Q-Excel file for data entry
}
