% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_planning_details}
\alias{load_planning_details}
\title{Loads Excel sheet "planning_area_details"}
\usage{
load_planning_details(data.dir, filename)
}
\arguments{
\item{data.dir}{The path of the entry data table.}

\item{filename}{Name of the R2Q-Excel File including ".xlsx".}
}
\value{
The Excel sheet as data frame
}
\description{
This functions loads the data from the sheet "pollution_data" within the
R2Q-Excel file for data entry
}
