% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_site_data}
\alias{load_site_data}
\title{Loading all information about local properties}
\usage{
load_site_data(data.dir, filename)
}
\arguments{
\item{data.dir}{the directory where the entry data table is}

\item{filename}{path of the site info Excel File including ".xlsx" .}
}
\value{
A list with all parameters from the site info table as seperate list
items. Per Parameter the item is a list containing the columnn names of the
site_info table
}
\description{
this functions loads the data from the sheet "site_data" within the data
entry excel file and returns the specified parameters in a list
}
