% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_site_info}
\alias{load_site_info}
\title{this functions either loads the Herne Example site data
(if nothing is specified) or a manually selected table with site infors}
\usage{
load_site_info(path = "inst/extdata/Eingabe/", filename = "site_info.csv")
}
\arguments{
\item{path}{path of the site info table.
The default value loads the data of example Herne}

\item{substances}{path of the site info table including ".csv" .
The default value loads the data of example Herne}
}
\value{
A list with all parameters from the site info table as seperate list
items. Per Parameter the item is a list containing the columnn names of the
site_info table
}
\description{
this functions either loads the Herne Example site data
(if nothing is specified) or a manually selected table with site infors
}
