% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_surface_data}
\alias{load_surface_data}
\title{Loading all details about planing area surface types}
\usage{
load_surface_data(data.dir, filename)
}
\arguments{
\item{data.dir}{the directory where the entry data table is}

\item{filename}{path of the site info Excel File including ".xlsx" .}
}
\value{
A data frame with all surface subtypes, the covered area in ha and the
corresponding runoff coefficient
}
\description{
this functions loads the data from the sheet "surface_data" within the data
entry excel file
}
