% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{massUnit_tranformation}
\alias{massUnit_tranformation}
\title{Transforms the mass units ng, ug, mg and g}
\usage{
massUnit_tranformation(original_unit, change)
}
\arguments{
\item{original_unit}{One of ng, ug ("u" instead of my), mg and g}

\item{change}{Integers between -2 and 2. Each integer represents a factor of
1000}
}
\value{
Character value of the transformed unit
}
\description{
Transforms the mass units ng, ug, mg and g
}
