% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{max_area}
\alias{max_area}
\title{old function, use max_area_steady_state or max_area_dynamic instead}
\usage{
max_area(Q_river, C_river, C_threshold, C_storm, coeff_runoff, rain)
}
\arguments{
\item{Q_river}{River flow in m3/time (time must represent impact duration)}

\item{C_river}{background concentration in mg/L}

\item{C_threshold}{concentration threshold in mg/L}

\item{C_storm}{concentration in stormwater in mg/L}

\item{coeff_runoff}{runoff coefficient of connected impervious area}

\item{rain}{rain amount in mm/time (time must represent impact duration)}
}
\value{
maximal connected impervious area in km2
}
\description{
based on concentration thresholds
based on DWA A102 approach
can be used for annual and acute impacts
if steady state concentration can be assumed or if inlets are
at one inlet point on the river stretch
}
