% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{max_area_dynamic}
\alias{max_area_dynamic}
\title{calculate maximal allowable connected area in a river catchment for a river section}
\usage{
max_area_dynamic(
  Q_river,
  C_river,
  C_threshold,
  C_storm,
  coeff_runoff,
  rain,
  delat_t,
  river_length,
  river_cross_section
)
}
\arguments{
\item{Q_river}{River flow in m3/time (time must represent impact duration)}

\item{C_river}{background concentration in mg/L}

\item{C_threshold}{concentration threshold in mg/L}

\item{C_storm}{concentration in stormwater in mg/L}

\item{coeff_runoff}{runoff coefficient of connected impervious area}

\item{rain}{rain amount in mm/time (time must represent impact duration)}

\item{river_length}{length of impacted urban river stretch in m}

\item{river_cross_section}{average cross section of river in m2}

\item{delta_t}{impact duration (typically = flow time in river stretch = rain duration)}
}
\value{
maximal connected impervious area in km2
}
\description{
based on concentration thresholds
based on DWA A102 approach (as far as applicable)
suggested for acute impacts if steady state concentration can not be assumed
, i.e. if inlets are distributed along the river stretch
}
