% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{max_area_steady_state}
\alias{max_area_steady_state}
\title{calculate maximal allowable connected area in a river catchment at steady state}
\usage{
max_area_steady_state(
  Q_river,
  Ci_river,
  Ci_threshold,
  Ci_storm,
  coeff_runoff,
  Q_rain
)
}
\arguments{
\item{Q_river}{Annual river flow in m³/a}

\item{Ci_river}{Background concentration for substance i. Concentration unit
must fit to Ci_threshold and Ci_storm.}

\item{Ci_threshold}{Threshold value for substance i. Concentration unit
must fit to Ci_river and Ci_storm.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i. Concentration unit
must fit to Ci_threshold and Ci_river.}

\item{coeff_runoff}{Run-off coefficient of connected impervious area}

\item{Q_rain}{Annual amount of rain amount in mm/a}
}
\value{
maximal connected impervious area in km2
}
\description{
based on concentration thresholds
based on DWA A102 approach
can be used for annual and acute impacts
if steady state concentration can be assumed or if inlets are
at one inlet point on the river stretch
}
