% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_assessments.R
\name{maxArea_by_pollution}
\alias{maxArea_by_pollution}
\title{Maximal connectable impervious area in the urabnised catchment and planning area}
\usage{
maxArea_by_pollution(combined_concentration_table, site_data, q_rain, t_rain)
}
\arguments{
\item{combined_concentration_table}{The table created by function
"combine_concentration_tables"}

\item{site_data}{The site specific data loaded with function "load_site_data"}

\item{q_rain}{Average rain intensity in L/(s*ha)}

\item{t_rain}{Duration of the rain event in s}
}
\value{
the combined concentration table is extend by the columns
connectable_catch: the tolerable area connected to the river under the given
area type mix in the whole catchment
connected_catch: the status quo are connected to the river under the given
area type mix in the whole catchment
connected_catch_ex: the status quo are connected to the river under the given
area type mix in the whole catchment without the planning area
connectable_planning_ha: the tolerable area connected to the river in ha under
the given area type mix in the planning area
connectable_planning_percent: the tolerable area connected to the river in
percent of total planning area under the given area type mix
}
\description{
This function calculates the maximal acceptabale impervious
area per substance depending on the threshold value type either by dynamic
or steady state concentration modeling
}
