% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{maxArea_event}
\alias{maxArea_event}
\title{Calculate connectable area to a river based on pollutant input within a
heavy rain event}
\usage{
maxArea_event(
  Q_river,
  Ci_river,
  Ci_threshold,
  Ci_storm,
  coeff_runoff,
  q_rain,
  t_rain,
  river_length,
  river_cross_section,
  catchment_area = 100
)
}
\arguments{
\item{Q_river}{Average River flow in m3/s}

\item{Ci_river}{Background concentration for substance i. Concentration unit
must fit to Ci_threshold and Ci_storm.}

\item{Ci_threshold}{Threshold value for substance i. Concentration unit
must fit to Ci_river and Ci_storm.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i. Concentration unit
must fit to Ci_threshold and Ci_river.}

\item{coeff_runoff}{runoff coefficient of connected impervious area}

\item{q_rain}{rain amount in mm/(ha*s)}

\item{t_rain}{duration of rain in s}

\item{river_length}{length of impacted urban river stretch in m}

\item{river_cross_section}{average cross section of river in m2}

\item{catchment_area}{Catchment area in ha.}
}
\value{
maximal connectable area in ha
}
\description{
Calculate connectable area to a river based on pollutant input within a
heavy rain event
}
\details{
The catchment_area is used as initial value for the optimisation algorithm.
The default 100 ha should be sufficient for most problems. In that case
the optimal solution between 0 and 1 000 km²
}
