% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{maxArea_year}
\alias{maxArea_year}
\title{Calculate connectable area in a river catchment based on a yearly regulated
Substance}
\usage{
maxArea_year(load_max, Ci_threshold, Ci_storm, coeff_runoff, Q_rain)
}
\arguments{
\item{load_max}{Annual maximal input of substance i. Mass unit corresponds
to concentration mass unit (two classes higher. i.e. concentration in mg/L
-> load in kg/a.}

\item{Ci_threshold}{Threshold value for substance i. Concentration unit
must fit to Ci_river and Ci_storm.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i. Concentration unit
must fit to Ci_threshold and Ci_river.}

\item{coeff_runoff}{Run-off coefficient of connected impervious area}

\item{Q_rain}{Annual amount of rain amount in mm/a}
}
\value{
maximal connectable area in ha
}
\description{
Calculate connectable area in a river catchment based on a yearly regulated
Substance
}
