% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{maxInput_year}
\alias{maxInput_year}
\title{Calculate maximal yearly pollutant input}
\usage{
maxInput_year(Q_river, Ci_river, Ci_storm, Ci_threshold)
}
\arguments{
\item{Q_river}{Annual river flow in m³/s}

\item{Ci_river}{Background concentration for substance i. Concentration unit
must fit to Ci_threshold.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i. Concentration unit
must fit to Ci_threshold and Ci_river.}

\item{Ci_threshold}{Threshold value for substance i. Concentration unit
must fit to Ci_river.}
}
\value{
Maximum tolerable pollutant input in mass per year. The mass unit
depends on the concentrations mass unit. It is transformed by 2 units.
i.e. concentration in ug/L -> load in g/a or concentration in mg/L ->
load in kg/a
}
\description{
Calculate maximal yearly pollutant input
}
