% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_assessments.R
\name{maxLoad_pollution}
\alias{maxLoad_pollution}
\title{Maximal load from the urbanised area and planning area}
\usage{
maxLoad_pollution(maxArea_list, site_data, rain)
}
\arguments{
\item{maxArea_list}{A list with maximal connectable areas created with
function "maxArea_by_pollution"}

\item{site_data}{The site specific data loaded with function "load_site_data"}

\item{rain}{Rain characteristics obtained with function "get_rain"}
}
\value{
A list with 4 data frames. 1) Information about the input data, 2) the
maximal tolerable pollutant load from the defined urban area, 3) the
maximal tolerable pollutant load from the planning area (area proportional)
and 4) maximal tolerable pollutant load from the planning area (given the
characterstics of the surounding urban area)
}
\description{
Maximal load from the urbanised area and planning area
}
