% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{merge_by_pollutant}
\alias{merge_by_pollutant}
\title{How to merge two R2Q concentration tables}
\usage{
merge_by_pollutant(dataFrame1, dataFrame2)
}
\arguments{
\item{dataFrame1}{A R2Q concentration data Frame}

\item{dataFrame2}{Another R2Q concentration data Fram}
}
\value{
A dataframe containing all substance measured in all units listed in
dataFrame1 and dataFrame2
}
\description{
This is a helping function for "combine_concentration_tables"
}
