% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{mixed_reactor_C}
\alias{mixed_reactor_C}
\title{Calculate the dynamic concentration in a river stretch}
\usage{
mixed_reactor_C(
  Q_river,
  Ci_river,
  Ci_storm,
  coeff_runoff,
  q_rain,
  t_rain,
  Area,
  V_river
)
}
\arguments{
\item{Q_river}{Average flow of the river in m³/s}

\item{Ci_river}{Background concentration for substance i. Concentration unit
must fit to Ci_threshold and Ci_storm.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i. Concentration unit
must fit to Ci_threshold and Ci_river.}

\item{coeff_runoff}{runoff coefficient of connected impervious area}

\item{q_rain}{Amount of rain amount in L/(s*ha)}

\item{t_rain}{duration of the rain in seconds}

\item{Area}{impervious, connected area in ha}

\item{V_river}{volume of the river in m³}
}
\value{
dynamic concentration after time t in the unit of the input
concentrations
}
\description{
based on mixed reactor approach
}
