% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planning_details.R
\name{planning_area_discharge}
\alias{planning_area_discharge}
\title{Calculates the pollutants load from the planning area into the surface water}
\usage{
planning_area_discharge(planning_data, q_rain, t_rain)
}
\arguments{
\item{planning_data}{The Excel sheet "planning_area_details" loaded by
\code{\link[=load_planning_details]{load_planning_details()}}}

\item{q_rain}{Rain intensity in L/(ha*s)}

\item{t_rain}{Rain length in s}
}
\value{
Pollutant load per event or per year, depending on the threshold value
definition. The mass unit is either mg or ug, depending on the input
concentration unit.
}
\description{
This functions reads specific runoff concentrations provided within the
package. For all substances with concentration data, the overall
discharged amount is calculated using the detailed information about
area functions within the planning area (Excel sheet:
"planning_area_details")
}
