% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_hazards}
\alias{plot_hazards}
\title{Plots the output of function \code{\link[=check_all_substances]{check_all_substances()}}}
\usage{
plot_hazards(
  hazards,
  title = "",
  xlabels = names(hazards),
  ylabels = names(hazards[[1]])
)
}
\arguments{
\item{hazards}{List created by \code{\link[=check_all_substances]{check_all_substances()}}}

\item{title}{Optional title}

\item{xlabels}{Character vector of the same length as Hazards to manually
enter x axis labels (-> Substance names)}

\item{ylabels}{Character vector of the same length as one vector of the
Hazards list entry to manually enter y axis labels (-> Landuse names)}
}
\value{
Plot with coloured rectangles representing the three different output
options from function \code{\link[=check_all_substances]{check_all_substances()}}
}
\description{
Plots the output of function \code{\link[=check_all_substances]{check_all_substances()}}
}
