% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2q_pal}
\alias{r2q_pal}
\title{Color palette for R2Q Plots}
\format{
Data frame with 3 columns for the colors and 6 rows for the shades
}
\source{
Logo of the R2Q project
}
\usage{
r2q_pal
}
\description{
Contains 3 different colors "orange", "blue" and "green" in 6 different
shades
}
\keyword{datasets}
