% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run immission-based tool}
\usage{
run_scenario(status_quo_list, scenario_name)
}
\arguments{
\item{status_quo_list}{A list created by}

\item{scenario_name}{The name of the excel sheet describing the planning.
THe excel file is the one that is defined by by the status_quo_list}
}
\value{
Saves all the output (figures and tables) in a folder created within
the file path. Furthermore, a table of maximum pollutant loads is returned
that can be used for evaluation of scenarios
}
\description{
Run immission-based tool
}
