% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_status_quo.R
\name{run_status_quo}
\alias{run_status_quo}
\title{Run immission-based tool}
\usage{
run_status_quo(path, filename, c_type)
}
\arguments{
\item{path}{File path of R2Q-Excel}

\item{filename}{File name of R2Q Excel (incuding .xlsx)}

\item{c_type}{A character defining the type of pollutant concentration in runoff
water. Either "average" for the median or "worstcase" for the 95th quantile.}
}
\value{
Saves all the output (figures and tables) in a folder created within
the file path. Furthermore, a table of maximum pollutant loads is returned
that can be used for evaluation of scenarios
}
\description{
Run immission-based tool
}
