% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{steady_state_C}
\alias{steady_state_C}
\title{calculate steady state concentration in river for complete mixing}
\usage{
steady_state_C(Q_river, C_river, C_storm, coeff_runoff, rain, Area)
}
\arguments{
\item{Q_river}{River flow in m3/time (time must represent impact duration)}

\item{C_river}{background concentration in mg/L}

\item{C_storm}{concentration in stormwater in mg/L}

\item{coeff_runoff}{runoff coefficient of connected impervious area}

\item{rain}{rain amount in mm/time (time must represent impact duration)}

\item{Area}{impervious, connected area in m2}
}
\value{
steady state concentration in mg/L
}
\description{
expected concentration at one inlet for complete mixing
or in one river stretch at steady state
can be used for annual and acute impacts
}
