% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{sub_id_to_name}
\alias{sub_id_to_name}
\title{Substance IDs within the package data tables are turned to substance names}
\usage{
sub_id_to_name(c_table)
}
\value{
c_table expanded by the column "substance"
}
\description{
Substance IDs within the package data tables are turned to substance names
}
\details{
This function ensures that every defined substance ID is part of the
concentration table, and every ID of the table is defined in the package
}
