% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{sub_OgRe_to_name}
\alias{sub_OgRe_to_name}
\title{OgRe substance names are turned to substance names used in the tables}
\usage{
sub_OgRe_to_name(c_table, all_substances = TRUE)
}
\arguments{
\item{c_table}{Data frame columne "Substance" or "substance" containing
OgRe substance names as defined in the OgRe data set.}

\item{all_substances}{If TRUE, all substances named in c_table or
in the substance ID table are kept. A warning is given if concentrations or
substance definitions are missing.}
}
\value{
c_table expanded by the column "substance"
}
\description{
OgRe substance names are turned to substance names used in the tables
}
\details{
To get an overiew of all Substance names, run function \code{\link[=get_subID]{get_subID()}}
}
