% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{substr_reverse}
\alias{substr_reverse}
\title{Start counting from the string end to get a substring}
\usage{
substr_reverse(x, rev_start, rev_stop, keep = TRUE)
}
\arguments{
\item{x}{Character String}

\item{rev_start, rev_stop}{The first and last value counted from the end of
the String. rev_stop > rev_start}

\item{keep}{If TRUE (default) the selection is return. Otherwise everything
but the selection is returend}
}
\value{
Character Vector
}
\description{
Start counting from the string end to get a substring
}
