% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{traffic_adaption}
\alias{traffic_adaption}
\title{Reduce the proportion of one area type according to the traffic}
\usage{
traffic_adaption(initial_share, traffic = "default")
}
\arguments{
\item{initial_share}{Proportion between 0 and 1 for one type of area}

\item{traffic}{Can either be "high" or "very_high"}
}
\value{
Returns the updated propordtion of the area type.
}
\description{
Reduce the proportion of one area type according to the traffic
}
\details{
A standard amount of traffic is already included in the area types.
However, more traffics (-> more high-traffic streets) can be added. 5 \% and
10 \% of the proportion of the area type are substracted if the traffic is
specified as "high" and "very_high", respecitvely.
}
