% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_assessments.R
\name{add_critical_loads}
\alias{add_critical_loads}
\title{Critical Pollutant loads into the surface water}
\usage{
add_critical_loads(max_area_table, site_data)
}
\arguments{
\item{max_area_table}{Table created with}

\item{site_data}{The site specific data loaded with function "loda_site_data"}
}
\value{
the combined max_area table is extend by the columns
crit_load_g_event: the maximal tolerated load per rain event.
Calculated for acute threshold values
crit_load_kg_year: the maximal tolerated load per year.
Calculated for annual threshold values
}
\description{
This function adds the critical loads to a table with maximal acceptable
impervious area
}
