% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{calculate_surface_discharge}
\alias{calculate_surface_discharge}
\title{calculate surface discharge from planning area}
\usage{
calculate_surface_discharge(area = 1, fD = 0.9, q_rain)
}
\arguments{
\item{area}{Impervious area in ha}

\item{fD}{Runoff coefficient of impervious area}

\item{q_rain}{Amount of rain in L / (s*ha)}
}
\value{
A numeric: stormwater run-off from planning area in L/s
}
\description{
this function calculates the surface discharge either from single values of
impervious area and runoff coefficent or from the input table surface_data
}
