% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{calculate_tolerable_discharge}
\alias{calculate_tolerable_discharge}
\title{Calculate Tolerable Discharge}
\usage{
calculate_tolerable_discharge(
  hydrology = import_hydology_data(),
  verbose = TRUE
)
}
\arguments{
\item{hydrology}{hydrology (tibble as retrieved by \link{import_hydrology_data})}

\item{verbose}{if TRUE returns results as informative messages, If FALSE only return numeric value for planning area.}
}
\value{
returns tolerable discharge for the planning area based on the data in hydrology table
}
\description{
Calculate Tolerable Discharge
}
\examples{
calculate_tolerable_discharge(hydrology = hydrology_data, verbose = F)
calculate_tolerable_discharge(hydrology = hydrology_data, verbose = T)
}
