% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_allowed_area}
\alias{get_allowed_area}
\title{Get average permeability}
\usage{
get_allowed_area(f_DA = 0.7, q_zul, R_Spende = 150)
}
\arguments{
\item{f_DA}{f_DA}

\item{q_zul}{q_zul}

\item{R_Spende}{R_Spende}
}
\value{
????
}
\description{
Get average permeability
}
