% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_areaType_runoff}
\alias{get_areaType_runoff}
\title{This function loads the area type specific pollutant runoff concentration
obtained by the OgRe Dataset and multiplies it with the proportion of the
correspoding area type in the catchment.}
\usage{
get_areaType_runoff(
  residential_city = 40,
  residential_suburban = 40,
  commercial = 20,
  main_road = NULL
)
}
\arguments{
\item{residential_suburban, residential_city, commercial, main_road}{Proportions
of the landuse types in percent}
}
\value{
A dataframe with the columns "Substance", "unit", "Mean" which is the
median value and "Q95" which is the 95th quantile.
}
\description{
This function loads the area type specific pollutant runoff concentration
obtained by the OgRe Dataset and multiplies it with the proportion of the
correspoding area type in the catchment.
}
