% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_Hq1_pnat}
\alias{get_Hq1_pnat}
\title{Calculate natural runoff based on slope of landscape}
\usage{
get_Hq1_pnat(slope, area_catch)
}
\arguments{
\item{slope}{slope of the planning area (unit \%)}

\item{area_catch}{catchment area  (in km2)}
}
\value{
Once-in-a-year natural discharge flow of the catchment in L/(s*km2)
}
\description{
Calculate natural runoff based on slope of landscape
}
\examples{
get_Hq1_pnat(slope = 0.1, area_catch = 5.62)
}
