% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_rain.R
\name{get_HQ_time_interval}
\alias{get_HQ_time_interval}
\title{get_HQ_time_intterval}
\usage{
get_HQ_time_interval(
  area_catch,
  river_cross_section,
  river_length,
  river_flow,
  Hq_pnat1_catch = NULL,
  slope = 0.1
)
}
\arguments{
\item{area_catch}{The catchment area in km2}

\item{river_cross_section}{The average river cross section in the catchment
in m2}

\item{river_length}{The length of the affected urban river stretch in m}

\item{river_flow}{The Average river flow in m3/s}

\item{Hq_pnat1_catch}{the natural catchment discharge for a yearly rain event
in L/(s*km2). If NULL it will be estimated by slope and area of the catchment}

\item{slope}{Average slope of the catchment in \% (Default is 0.1)}
}
\value{
Travel time of naturally discharged water within the catchment in minutes
}
\description{
This function calculates the time in minutes that is needed for the water to
travel through the affected urban river stretch for a yearly rain event,
based on a natural catchment discharge
}
\details{
The natural catchment discharge is estimated based on the supplementary
information of DWA-A 102-3. Unlike the rain intensity of a yearly rain, the
estimated natural discharge is independent of the rain duration.
According to DWA-A 102-3 two more factors would increase the travel time that
are not considered here: 1) The longest travel time within the sewer network
of the planning area and 2) The increased water level leading to a higher
river cross section and thus to a longer travel time.
Instead a constant of 60 minutes is added to the calculated travel time. This
also ensures that the rain duration is high enough (> 60 min) for a
toxicological relevance.
}
