% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSTRA_timeseries.R
\name{get_KOSTRA}
\alias{get_KOSTRA}
\title{Get KOSTRA}
\usage{
get_KOSTRA(
  path = system.file("extdata/KOSTRA", package = "r2q"),
  coord_vector = c(3813634.44, 2753912.5),
  location_name = "Herne",
  base_name = "GIS_KOSTRA-DWD-2010R_D",
  duration_string = "0060",
  plot = TRUE
)
}
\arguments{
\item{path}{path to shapefiles (default: system.file("extdata/KOSTRA", package = "r2q"))}

\item{coord_vector}{coordinates in 3034}

\item{location_name}{default: "Herne"}

\item{base_name}{default: "GIS_KOSTRA-DWD-2010R_D"}

\item{duration_string}{default: "0060"}

\item{plot}{boolean (TRUE for plotting or FALSE if plotting is not required)}
}
\value{
Return numeric results (and also plot if parameter plot = TRUE)
}
\description{
Get KOSTRA
}
\examples{
  \dontrun{
  link_to_coord_help <- "https://epsg.io/transform#s_srs=4326&t_srs=3034"
  
  # Example default values Herne
  get_KOSTRA()
  
  # Example 10 min for Berlin
  get_KOSTRA(coord_vector = c(4217676.98, 2862423.69),
  duration_string = "0010",
  location_name = "Berlin", plot = TRUE)
  }
}
