% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_stormwater_concentrations}
\alias{get_stormwater_concentrations}
\title{get average concentrations in stormwater runoff}
\usage{
get_stormwater_concentrations(substances = NULL)
}
\arguments{
\item{substances}{substance name or vector of substance names in German
if left NULL all parameters measured in OgRe will be loaded}
}
\value{
data.frame with mean(x), mean (log x) and sd (log x)
}
\description{
assembles concentrations in stormwater runoff
for selected substances based on Berlin OgRe data set
from csv table. In addition to averages, coeffs for log-normal
distribution are assessed
}
