% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_x}
\alias{get_x}
\title{Calculate acceptable additional runoff factor x}
\usage{
get_x(Hq1_pnat, Hq2_pnat)
}
\arguments{
\item{Hq1_pnat}{potential annual natural discharge flow in L/(s*km2)}

\item{Hq2_pnat}{potential biennial natural discharge flow in L/(s*km2)}
}
\value{
dimensionless factor regulating tolerable additional anthropogenic discharge
}
\description{
Calculate acceptable additional runoff factor x
}
