% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{hydrology_assessment}
\alias{hydrology_assessment}
\title{Maximal connectable impervious area based on hydrologic conditions}
\usage{
hydrology_assessment(site_data, q_rain)
}
\arguments{
\item{site_data}{The site specific data loaded with function "loda_site_data"}

\item{q_rain}{characteristic rainfall in L/(s*ha)}
}
\value{
the combined max_area table is extend by a row with the result of the
hydolic assessment.
}
\description{
Maximal connectable impervious area based on hydrologic conditions
}
