% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_areaTypes}
\alias{load_areaTypes}
\title{Loading all details about catchment area types}
\usage{
load_areaTypes(data.dir, filename)
}
\arguments{
\item{data.dir}{The directory of the entry data table.}

\item{filename}{Name of the R2Q-Excel File including ".xlsx".}
}
\value{
A vector of length 5. Entries 1 to 4 describe the proportion of the area
types "residential_suburban", "residential_city", "industry" and (high-
traffic) "street". The proportion is referred only to the connected area.
The 5th value is the overall proportion of connected area.
}
\description{
this functions loads the data from the sheet "surface_areaType" within the data
entry excel file
}
\examples{
load_site_data(
data.dir = system.file("extdata/Data_entry", package = "r2q"), 
filename = "Baukau.xlsx")

}
