% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_info.R
\name{load_site_data}
\alias{load_site_data}
\title{Loading site specific information}
\usage{
load_site_data(data.dir, filename)
}
\arguments{
\item{data.dir}{The directory of the entry data table.}

\item{filename}{Name of the R2Q-Excel File including ".xlsx".}
}
\value{
A list with all parameters from the site info table as seperate list
items. Per Parameter the item is a list containing the columnn names of the
site_info table
}
\description{
this functions loads the data from the sheet "site_data" within the data
entry excel file and returns the specified parameters in a list
}
\examples{
load_site_data(
data.dir = system.file("extdata/Data_entry", package = "r2q"), 
filename = "Bsp_Herne.xlsx")

}
