% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{max_area_dynamic}
\alias{max_area_dynamic}
\title{calculate maximal allowable connected area in a river catchment for a river section}
\usage{
max_area_dynamic(
  Q_river,
  Ci_river,
  Ci_threshold,
  Ci_storm,
  coeff_runoff,
  q_rain,
  t_rain,
  river_length,
  river_cross_section,
  catchment_area = 100
)
}
\arguments{
\item{Q_river}{Average River flow in m3/s}

\item{Ci_river}{Background concentration for substance i. Concentration unit
must fit to Ci_threshold and Ci_storm.}

\item{Ci_threshold}{Threshold value for substance i. Concentration unit
must fit to Ci_river and Ci_storm.}

\item{Ci_storm}{Concentration in stormwater run-off for substance i. Concentration unit
must fit to Ci_threshold and Ci_river.}

\item{coeff_runoff}{runoff coefficient of connected impervious area}

\item{q_rain}{rain amount in mm/(ha*s)}

\item{t_rain}{duration of rain in s}

\item{river_length}{length of impacted urban river stretch in m}

\item{river_cross_section}{average cross section of river in m2}

\item{catchment_area}{Catchment area in ha.}
}
\value{
maximal connected impervious area in ha
}
\description{
based on concentration thresholds
based on DWA A102 approach (as far as applicable)
suggested for acute impacts if steady state concentration can not be assumed
, i.e. if inlets are distributed along the river stretch
}
\details{
The catchment_area is used as initial value for the optimisation algorithm.
The default 100 ha should be sufficient for most problems. In that case
the optimal solution between 0 and 1 000 km²
}
