% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_area.R
\name{mixed_reactor_C}
\alias{mixed_reactor_C}
\title{calculate dynamic concentration in river stretch}
\usage{
mixed_reactor_C(
  Q_river,
  C_river,
  C_threshold,
  C_storm,
  coeff_runoff,
  rain,
  delta_t,
  Area,
  Vol
)
}
\arguments{
\item{Q_river}{River flow in m3/time (time must represent impact duration)}

\item{C_river}{background concentration in mg/L}

\item{C_threshold}{concentration threshold in mg/L}

\item{C_storm}{concentration in stormwater in mg/L}

\item{coeff_runoff}{runoff coefficient of connected impervious area}

\item{rain}{rain amount in mm/time (time must represent impact duration)}

\item{delta_t}{impact duration (typically = flow time in river stretch = rain duration, same time unit as rain/flow)}

\item{Area}{impervious, connected area in m2}

\item{Vol}{reactor volume (= river stretch volume) in m3}
}
\value{
dynamic concentration after time delta_t in mg/L
}
\description{
based on mixed reactor approach
}
