% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_connectable_urban_area}
\alias{plot_connectable_urban_area}
\title{Plot of relative connectable area}
\usage{
plot_connectable_urban_area(r2q_substance, r2q_hydrology = NULL)
}
\arguments{
\item{r2q_substance}{Assessment output created by \code{\link[=assess_all_hazards]{assess_all_hazards()}}}

\item{r2q_hydrology}{Assessment output created by \code{\link[=hydrology_assessment]{hydrology_assessment()}}.
Is NULL by default, so that the plot can be created for substances only}
}
\description{
Share of the urban area that can be connected to the seperate sewer system
without exceeding the threshold values (and without further treatment)
}
