% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{sub_id_to_name}
\alias{sub_id_to_name}
\title{Substance IDs within the package data tables are turned to substance names}
\usage{
sub_id_to_name(c_table, all_substances = TRUE)
}
\arguments{
\item{c_table}{Data frame with column "Substance" or "substance" containing
substance IDs as defined in the package substance ID table \code{\link[=get_subID]{get_subID()}}}

\item{all_substances}{If TRUE, all substances named in c_table or
in the substance ID table are kept. A warning is given if concentrations or
substance definitions are missing.}
}
\value{

}
\description{
Substance IDs within the package data tables are turned to substance names
}
