% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{sub_OgRe_to_name}
\alias{sub_OgRe_to_name}
\title{OgRe substance names are turned to substance names used in the tables}
\usage{
sub_OgRe_to_name(c_table)
}
\value{
c_table expanded by the column "substance"
}
\description{
OgRe substance names are turned to substance names used in the tables
}
\details{
This function ensures that every defined substance is part of the
concentration table, and every OgRe Substance of the table is defined in the package
}
